Refactor app.py: Added imports for embedding and feedback services, and commented out function definitions for code extraction and analysis. Updated process_qa endpoint to accept request parameter.
