+++ app.py
@@ -1,6 +1,7 @@
-import ast
 import zipfile
 import os
+import ast
+import requests
 import numpy as np
 from fastapi import FastAPI, UploadFile, File, Form, Request
 from fastapi.responses import HTMLResponse
@@ -26,6 +27,11 @@ import urllib.parse
 import shutil
 from fastapi.responses import JSONResponse
 from utils.prompts import OPENAI_CODE_FEEDBACK_PROMPT
+from utils.embedding_service import get_embedding, compute_similarity, get_embedding_ollama, get_embedding_openai
+from utils.feedback_service import generate_feedback    
+from utils.db_service import store_embedding, retrieve_similar_contexts
+from utils.code_analyzer import analyze_code_structure, generate_recommendations, extract_functions_from_file, extract_functions_from_zip
+
 
 # Configure logging first before using logger
 logging.basicConfig(
@@ -170,409 +176,409 @@ except Exception as e:
     logger.error(f"Error initializing TextRAGProcessor: {e}")
     text_rag_processor = None
 
-def extract_functions_from_file(file_path: str) -> Dict[str, str]:
-    """Extract function definitions from a Python file."""
-    try:
-        with open(file_path, 'r', encoding='utf-8') as file:
-            code = file.read()
-        tree = ast.parse(code)
-        functions = [node for node in tree.body if isinstance(node, ast.FunctionDef)]
-        return {func.name: ast.get_source_segment(code, func) for func in functions}
-    except Exception as e:
-        logger.warning(f"Error processing file {file_path}: {e}")
-        return {}
+# def extract_functions_from_file(file_path: str) -> Dict[str, str]:
+#     """Extract function definitions from a Python file."""
+#     try:
+#         with open(file_path, 'r', encoding='utf-8') as file:
+#             code = file.read()
+#         tree = ast.parse(code)
+#         functions = [node for node in tree.body if isinstance(node, ast.FunctionDef)]
+#         return {func.name: ast.get_source_segment(code, func) for func in functions}
+#     except Exception as e:
+#         logger.warning(f"Error processing file {file_path}: {e}")
+#         return {}
 
-def extract_functions_from_zip(zip_path: str) -> Dict[str, str]:
-    """Extract all functions from Python files in a zip."""
-    logger.info(f"Starting function extraction from ZIP: {os.path.basename(zip_path)}")
-    func_codes = {}
-    with tempfile.TemporaryDirectory() as extract_dir:
-        try:
-            # Validate zip file before extracting
-            if not zipfile.is_zipfile(zip_path):
-                logger.error(f"Invalid ZIP file: {zip_path}")
-                raise ValueError(f"The provided file is not a valid ZIP archive.")
+# def extract_functions_from_zip(zip_path: str) -> Dict[str, str]:
+#     """Extract all functions from Python files in a zip."""
+#     logger.info(f"Starting function extraction from ZIP: {os.path.basename(zip_path)}")
+#     func_codes = {}
+#     with tempfile.TemporaryDirectory() as extract_dir:
+#         try:
+#             # Validate zip file before extracting
+#             if not zipfile.is_zipfile(zip_path):
+#                 logger.error(f"Invalid ZIP file: {zip_path}")
+#                 raise ValueError(f"The provided file is not a valid ZIP archive.")
                 
-            with zipfile.ZipFile(zip_path, 'r') as zip_ref:
-                # Extract with careful error handling
-                try:
-                    logger.info(f"Extracting ZIP file to temporary directory: {extract_dir}")
-                    zip_ref.extractall(extract_dir)
-                    logger.info(f"Successfully extracted {len(zip_ref.namelist())} files from ZIP")
-                except zipfile.BadZipFile:
-                    logger.error(f"Bad ZIP file format: {zip_path}")
-                    raise ValueError(f"The ZIP file is corrupt or in an unsupported format.")
-                except PermissionError:
-                    logger.error(f"Permission denied when extracting: {zip_path}")
-                    raise ValueError(f"Permission denied when extracting the ZIP file.")
-                except Exception as extract_error:
-                    logger.error(f"Error extracting ZIP: {extract_error}")
-                    raise ValueError(f"Failed to extract the ZIP file: {str(extract_error)}")
+#             with zipfile.ZipFile(zip_path, 'r') as zip_ref:
+#                 # Extract with careful error handling
+#                 try:
+#                     logger.info(f"Extracting ZIP file to temporary directory: {extract_dir}")
+#                     zip_ref.extractall(extract_dir)
+#                     logger.info(f"Successfully extracted {len(zip_ref.namelist())} files from ZIP")
+#                 except zipfile.BadZipFile:
+#                     logger.error(f"Bad ZIP file format: {zip_path}")
+#                     raise ValueError(f"The ZIP file is corrupt or in an unsupported format.")
+#                 except PermissionError:
+#                     logger.error(f"Permission denied when extracting: {zip_path}")
+#                     raise ValueError(f"Permission denied when extracting the ZIP file.")
+#                 except Exception as extract_error:
+#                     logger.error(f"Error extracting ZIP: {extract_error}")
+#                     raise ValueError(f"Failed to extract the ZIP file: {str(extract_error)}")
             
-            # Check if any Python files were found
-            python_files_found = False
-            python_files_count = 0
+#             # Check if any Python files were found
+#             python_files_found = False
+#             python_files_count = 0
             
-            logger.info(f"Scanning directory for Python files: {extract_dir}")
-            for root, dirs, files in os.walk(extract_dir):
-                logger.debug(f"Scanning directory: {root} (contains {len(files)} files)")
-                for file in files:
-                    if file.endswith('.py'):
-                        python_files_count += 1
-                        python_files_found = True
-                        file_path = os.path.join(root, file)
-                        logger.info(f"Processing Python file: {os.path.relpath(file_path, extract_dir)}")
-                        try:
-                            file_functions = extract_functions_from_file(file_path)
-                            if file_functions:
-                                logger.info(f"Found {len(file_functions)} functions in {os.path.relpath(file_path, extract_dir)}")
-                                for func_name in file_functions:
-                                    logger.debug(f"  - Function: {func_name}")
-                            else:
-                                logger.info(f"No functions found in {os.path.relpath(file_path, extract_dir)}")
-                            func_codes.update(file_functions)
-                        except Exception as file_error:
-                            logger.warning(f"Skipping file {file_path}: {file_error}")
+#             logger.info(f"Scanning directory for Python files: {extract_dir}")
+#             for root, dirs, files in os.walk(extract_dir):
+#                 logger.debug(f"Scanning directory: {root} (contains {len(files)} files)")
+#                 for file in files:
+#                     if file.endswith('.py'):
+#                         python_files_count += 1
+#                         python_files_found = True
+#                         file_path = os.path.join(root, file)
+#                         logger.info(f"Processing Python file: {os.path.relpath(file_path, extract_dir)}")
+#                         try:
+#                             file_functions = extract_functions_from_file(file_path)
+#                             if file_functions:
+#                                 logger.info(f"Found {len(file_functions)} functions in {os.path.relpath(file_path, extract_dir)}")
+#                                 for func_name in file_functions:
+#                                     logger.debug(f"  - Function: {func_name}")
+#                             else:
+#                                 logger.info(f"No functions found in {os.path.relpath(file_path, extract_dir)}")
+#                             func_codes.update(file_functions)
+#                         except Exception as file_error:
+#                             logger.warning(f"Skipping file {file_path}: {file_error}")
             
-            logger.info(f"Found {python_files_count} Python files with {len(func_codes)} total functions")
+#             logger.info(f"Found {python_files_count} Python files with {len(func_codes)} total functions")
             
-            if not python_files_found:
-                logger.warning(f"No Python files found in ZIP: {zip_path}")
-                raise ValueError(f"No Python (.py) files found in the ZIP archive. Please check the contents of your submission.")
+#             if not python_files_found:
+#                 logger.warning(f"No Python files found in ZIP: {zip_path}")
+#                 raise ValueError(f"No Python (.py) files found in the ZIP archive. Please check the contents of your submission.")
                 
-        except ValueError as value_error:
-            # Re-raise ValueError with the message intact
-            raise value_error
-        except Exception as e:
-            logger.error(f"Unexpected error processing ZIP file {zip_path}: {e}")
-            raise ValueError(f"Error processing ZIP file: {str(e)}")
+#         except ValueError as value_error:
+#             # Re-raise ValueError with the message intact
+#             raise value_error
+#         except Exception as e:
+#             logger.error(f"Unexpected error processing ZIP file {zip_path}: {e}")
+#             raise ValueError(f"Error processing ZIP file: {str(e)}")
     
-    if not func_codes:
-        logger.warning(f"No functions found in any Python files from ZIP: {zip_path}")
-        raise ValueError(f"No Python functions were found in the submitted files. Please ensure your code contains function definitions.")
+#     if not func_codes:
+#         logger.warning(f"No functions found in any Python files from ZIP: {zip_path}")
+#         raise ValueError(f"No Python functions were found in the submitted files. Please ensure your code contains function definitions.")
     
-    logger.info(f"Successfully extracted {len(func_codes)} functions: {', '.join(func_codes.keys())}")    
-    return func_codes
+#     logger.info(f"Successfully extracted {len(func_codes)} functions: {', '.join(func_codes.keys())}")    
+#     return func_codes
 
-def get_embedding_ollama(text: str) -> np.ndarray:
-    """Generate embedding using Ollama's local model."""
-    try:
-        response = requests.post(
-            OLLAMA_API_URL,
-            json={"model": "llama3.2:3b", "prompt": text},
-            timeout=30
-        )
-        response.raise_for_status()
-        result = response.json()
-        if "embedding" not in result:
-            raise ValueError("No embedding in Ollama response")
-        return np.array(result["embedding"])
-    except Exception as e:
-        logger.error(f"Ollama embedding generation failed: {e}")
-        raise ValueError(f"Ollama embedding generation failed: {str(e)}")
+# def get_embedding_ollama(text: str) -> np.ndarray:
+#     """Generate embedding using Ollama's local model."""
+#     try:
+#         response = requests.post(
+#             OLLAMA_API_URL,
+#             json={"model": "llama3.2:3b", "prompt": text},
+#             timeout=30
+#         )
+#         response.raise_for_status()
+#         result = response.json()
+#         if "embedding" not in result:
+#             raise ValueError("No embedding in Ollama response")
+#         return np.array(result["embedding"])
+#     except Exception as e:
+#         logger.error(f"Ollama embedding generation failed: {e}")
+#         raise ValueError(f"Ollama embedding generation failed: {str(e)}")
 
-def get_embedding_openai(text: str) -> np.ndarray:
-    """Generate embedding using OpenAI's API."""
-    if not OPENAI_API_KEY:
-        raise ValueError("OpenAI API key not configured")
+# def get_embedding_openai(text: str) -> np.ndarray:
+#     """Generate embedding using OpenAI's API."""
+#     if not OPENAI_API_KEY:
+#         raise ValueError("OpenAI API key not configured")
     
-    try:
-        response = requests.post(
-            OPENAI_API_URL,
-            headers={
-                "Authorization": f"Bearer {OPENAI_API_KEY}",
-                "Content-Type": "application/json"
-            },
-            json={
-                "model": "text-embedding-ada-002",
-                "input": text
-            },
-            timeout=10
-        )
-        response.raise_for_status()
-        return np.array(response.json()["data"][0]["embedding"])
-    except Exception as e:
-        logger.error(f"OpenAI embedding generation failed: {e}")
-        raise ValueError(f"OpenAI embedding generation failed: {str(e)}")
+#     try:
+#         response = requests.post(
+#             OPENAI_API_URL,
+#             headers={
+#                 "Authorization": f"Bearer {OPENAI_API_KEY}",
+#                 "Content-Type": "application/json"
+#             },
+#             json={
+#                 "model": "text-embedding-ada-002",
+#                 "input": text
+#             },
+#             timeout=10
+#         )
+#         response.raise_for_status()
+#         return np.array(response.json()["data"][0]["embedding"])
+#     except Exception as e:
+#         logger.error(f"OpenAI embedding generation failed: {e}")
+#         raise ValueError(f"OpenAI embedding generation failed: {str(e)}")
 
-def get_embedding(text: str, model: EmbeddingModel = EmbeddingModel.OLLAMA) -> np.ndarray:
-    """Generate embedding based on selected model."""
-    logger.debug(f"Generating embedding using model: {model}")
-    start_time = datetime.now()
+# def get_embedding(text: str, model: EmbeddingModel = EmbeddingModel.OLLAMA) -> np.ndarray:
+#     """Generate embedding based on selected model."""
+#     logger.debug(f"Generating embedding using model: {model}")
+#     start_time = datetime.now()
     
-    try:
-        if model == EmbeddingModel.OLLAMA:
-            embedding = get_embedding_ollama(text)
-        else:
-            embedding = get_embedding_openai(text)
+#     try:
+#         if model == EmbeddingModel.OLLAMA:
+#             embedding = get_embedding_ollama(text)
+#         else:
+#             embedding = get_embedding_openai(text)
         
-        elapsed = (datetime.now() - start_time).total_seconds()
-        logger.debug(f"Generated embedding successfully in {elapsed:.2f}s ({len(text)} chars, model: {model})")
-        return embedding
-    except Exception as e:
-        elapsed = (datetime.now() - start_time).total_seconds()
-        logger.error(f"Embedding generation failed after {elapsed:.2f}s: {e}")
-        raise
+#         elapsed = (datetime.now() - start_time).total_seconds()
+#         logger.debug(f"Generated embedding successfully in {elapsed:.2f}s ({len(text)} chars, model: {model})")
+#         return embedding
+#     except Exception as e:
+#         elapsed = (datetime.now() - start_time).total_seconds()
+#         logger.error(f"Embedding generation failed after {elapsed:.2f}s: {e}")
+#         raise
 
-def compute_similarity(emb1: np.ndarray, emb2: np.ndarray) -> float:
-    """Compute cosine similarity between two embeddings."""
-    try:
-        return float(1 - cosine(emb1, emb2))
-    except Exception as e:
-        logger.error(f"Error computing similarity: {e}")
-        return 0.0
+# def compute_similarity(emb1: np.ndarray, emb2: np.ndarray) -> float:
+#     """Compute cosine similarity between two embeddings."""
+#     try:
+#         return float(1 - cosine(emb1, emb2))
+#     except Exception as e:
+#         logger.error(f"Error computing similarity: {e}")
+#         return 0.0
 
-def analyze_code_structure(student_code: str, ideal_code: str) -> Dict[str, Any]:
-    """Analyze and compare code structure between student and ideal solutions."""
-    logger.info("Analyzing and comparing code structure")
-    start_time = datetime.now()
+# def analyze_code_structure(student_code: str, ideal_code: str) -> Dict[str, Any]:
+#     """Analyze and compare code structure between student and ideal solutions."""
+#     logger.info("Analyzing and comparing code structure")
+#     start_time = datetime.now()
     
-    try:
-        # Parse both code snippets
-        logger.debug("Parsing student and ideal code with ast")
-        student_tree = ast.parse(student_code)
-        ideal_tree = ast.parse(ideal_code)
+#     try:
+#         # Parse both code snippets
+#         logger.debug("Parsing student and ideal code with ast")
+#         student_tree = ast.parse(student_code)
+#         ideal_tree = ast.parse(ideal_code)
         
-        # Extract variables from both
-        logger.debug("Extracting variables")
-        student_vars = set()
-        ideal_vars = set()
-        for node in ast.walk(student_tree):
-            if isinstance(node, ast.Name) and isinstance(node.ctx, ast.Store):
-                student_vars.add(node.id)
-        for node in ast.walk(ideal_tree):
-            if isinstance(node, ast.Name) and isinstance(node.ctx, ast.Store):
-                ideal_vars.add(node.id)
+#         # Extract variables from both
+#         logger.debug("Extracting variables")
+#         student_vars = set()
+#         ideal_vars = set()
+#         for node in ast.walk(student_tree):
+#             if isinstance(node, ast.Name) and isinstance(node.ctx, ast.Store):
+#                 student_vars.add(node.id)
+#         for node in ast.walk(ideal_tree):
+#             if isinstance(node, ast.Name) and isinstance(node.ctx, ast.Store):
+#                 ideal_vars.add(node.id)
         
-        # Extract control flow structures
-        logger.debug("Extracting control flow structures")
-        student_ctrl = []
-        ideal_ctrl = []
-        for node in ast.walk(student_tree):
-            if isinstance(node, (ast.If, ast.For, ast.While)):
-                student_ctrl.append(type(node).__name__)
-        for node in ast.walk(ideal_tree):
-            if isinstance(node, (ast.If, ast.For, ast.While)):
-                ideal_ctrl.append(type(node).__name__)
+#         # Extract control flow structures
+#         logger.debug("Extracting control flow structures")
+#         student_ctrl = []
+#         ideal_ctrl = []
+#         for node in ast.walk(student_tree):
+#             if isinstance(node, (ast.If, ast.For, ast.While)):
+#                 student_ctrl.append(type(node).__name__)
+#         for node in ast.walk(ideal_tree):
+#             if isinstance(node, (ast.If, ast.For, ast.While)):
+#                 ideal_ctrl.append(type(node).__name__)
         
-        # Extract function calls
-        logger.debug("Extracting function calls")
-        student_calls = set()
-        ideal_calls = set()
-        for node in ast.walk(student_tree):
-            if isinstance(node, ast.Call) and hasattr(node.func, 'id'):
-                student_calls.add(node.func.id)
-        for node in ast.walk(ideal_tree):
-            if isinstance(node, ast.Call) and hasattr(node.func, 'id'):
-                ideal_calls.add(node.func.id)
+#         # Extract function calls
+#         logger.debug("Extracting function calls")
+#         student_calls = set()
+#         ideal_calls = set()
+#         for node in ast.walk(student_tree):
+#             if isinstance(node, ast.Call) and hasattr(node.func, 'id'):
+#                 student_calls.add(node.func.id)
+#         for node in ast.walk(ideal_tree):
+#             if isinstance(node, ast.Call) and hasattr(node.func, 'id'):
+#                 ideal_calls.add(node.func.id)
         
-        # Compare and return differences
-        result = {
-            "variables": {
-                "missing_variables": list(ideal_vars - student_vars),
-                "extra_variables": list(student_vars - ideal_vars)
-            },
-            "control_flow": {
-                "missing_control_structures": [c for c in ideal_ctrl if c not in student_ctrl],
-                "extra_control_structures": [c for c in student_ctrl if c not in ideal_ctrl]
-            },
-            "function_calls": {
-                "missing_calls": list(ideal_calls - student_calls),
-                "extra_calls": list(student_calls - ideal_calls)
-            }
-        }
+#         # Compare and return differences
+#         result = {
+#             "variables": {
+#                 "missing_variables": list(ideal_vars - student_vars),
+#                 "extra_variables": list(student_vars - ideal_vars)
+#             },
+#             "control_flow": {
+#                 "missing_control_structures": [c for c in ideal_ctrl if c not in student_ctrl],
+#                 "extra_control_structures": [c for c in student_ctrl if c not in ideal_ctrl]
+#             },
+#             "function_calls": {
+#                 "missing_calls": list(ideal_calls - student_calls),
+#                 "extra_calls": list(student_calls - ideal_calls)
+#             }
+#         }
         
-        elapsed = (datetime.now() - start_time).total_seconds()
-        logger.info(f"Code structure analysis completed in {elapsed:.2f}s")
-        logger.debug(f"Analysis results: {json.dumps(result, indent=2)}")
+#         elapsed = (datetime.now() - start_time).total_seconds()
+#         logger.info(f"Code structure analysis completed in {elapsed:.2f}s")
+#         logger.debug(f"Analysis results: {json.dumps(result, indent=2)}")
         
-        return result
-    except Exception as e:
-        elapsed = (datetime.now() - start_time).total_seconds()
-        logger.error(f"Error analyzing code structure after {elapsed:.2f}s: {e}")
-        return {
-            "variables": {"missing_variables": [], "extra_variables": []},
-            "control_flow": {"missing_control_structures": [], "extra_control_structures": []},
-            "function_calls": {"missing_calls": [], "extra_calls": []}
-        }
+#         return result
+#     except Exception as e:
+#         elapsed = (datetime.now() - start_time).total_seconds()
+#         logger.error(f"Error analyzing code structure after {elapsed:.2f}s: {e}")
+#         return {
+#             "variables": {"missing_variables": [], "extra_variables": []},
+#             "control_flow": {"missing_control_structures": [], "extra_control_structures": []},
+#             "function_calls": {"missing_calls": [], "extra_calls": []}
+#         }
 
-def generate_recommendations(structure_analysis: Dict[str, Any]) -> List[str]:
-    """Generate recommendations based on structure analysis."""
-    recommendations = []
+# def generate_recommendations(structure_analysis: Dict[str, Any]) -> List[str]:
+#     """Generate recommendations based on structure analysis."""
+#     recommendations = []
     
-    # Variable recommendations
-    missing_vars = structure_analysis["variables"]["missing_variables"]
-    if missing_vars:
-        recommendations.append(f"Add missing variables: {', '.join(missing_vars)}")
+#     # Variable recommendations
+#     missing_vars = structure_analysis["variables"]["missing_variables"]
+#     if missing_vars:
+#         recommendations.append(f"Add missing variables: {', '.join(missing_vars)}")
     
-    extra_vars = structure_analysis["variables"]["extra_variables"]
-    if extra_vars:
-        recommendations.append(f"Consider removing unnecessary variables: {', '.join(extra_vars)}")
+#     extra_vars = structure_analysis["variables"]["extra_variables"]
+#     if extra_vars:
+#         recommendations.append(f"Consider removing unnecessary variables: {', '.join(extra_vars)}")
     
-    # Control flow recommendations
-    missing_ctrl = structure_analysis["control_flow"]["missing_control_structures"]
-    if missing_ctrl:
-        recommendations.append(f"Add missing control structures: {', '.join(missing_ctrl)}")
+#     # Control flow recommendations
+#     missing_ctrl = structure_analysis["control_flow"]["missing_control_structures"]
+#     if missing_ctrl:
+#         recommendations.append(f"Add missing control structures: {', '.join(missing_ctrl)}")
     
-    # Function call recommendations
-    missing_calls = structure_analysis["function_calls"]["missing_calls"]
-    if missing_calls:
-        recommendations.append(f"Add missing function calls: {', '.join(missing_calls)}")
+#     # Function call recommendations
+#     missing_calls = structure_analysis["function_calls"]["missing_calls"]
+#     if missing_calls:
+#         recommendations.append(f"Add missing function calls: {', '.join(missing_calls)}")
     
-    if not recommendations:
-        recommendations.append("Your implementation looks good structurally!")
+#     if not recommendations:
+#         recommendations.append("Your implementation looks good structurally!")
     
-    return recommendations
+#     return recommendations
 
-def store_embedding(func_name: str, code: str, embedding: np.ndarray) -> None:
-    """Store embedding in MongoDB."""
-    try:
-        if mongo_client is not None and embeddings_collection is not None:
-            embeddings_collection.insert_one({
-                "function_name": func_name,
-                "code": code,
-                "embedding": embedding.tolist(),
-                "timestamp": datetime.now()
-            })
-            logger.debug(f"Embedding stored for function: {func_name}")
-        else:
-            logger.warning(f"Embedding not stored for function {func_name} - MongoDB connection unavailable")
-    except Exception as e:
-        logger.error(f"Error storing embedding: {e}")
+# def store_embedding(func_name: str, code: str, embedding: np.ndarray) -> None:
+#     """Store embedding in MongoDB."""
+#     try:
+#         if mongo_client is not None and embeddings_collection is not None:
+#             embeddings_collection.insert_one({
+#                 "function_name": func_name,
+#                 "code": code,
+#                 "embedding": embedding.tolist(),
+#                 "timestamp": datetime.now()
+#             })
+#             logger.debug(f"Embedding stored for function: {func_name}")
+#         else:
+#             logger.warning(f"Embedding not stored for function {func_name} - MongoDB connection unavailable")
+#     except Exception as e:
+#         logger.error(f"Error storing embedding: {e}")
 
-def retrieve_similar_contexts(query_embedding: np.ndarray, top_k: int = 9) -> List[Dict[str, Any]]:
-    """Retrieve similar code contexts based on embedding similarity."""
-    logger.info(f"Retrieving similar contexts (top_{top_k})")
-    start_time = datetime.now()
+# def retrieve_similar_contexts(query_embedding: np.ndarray, top_k: int = 9) -> List[Dict[str, Any]]:
+#     """Retrieve similar code contexts based on embedding similarity."""
+#     logger.info(f"Retrieving similar contexts (top_{top_k})")
+#     start_time = datetime.now()
     
-    try:
-        if mongo_client is not None and embeddings_collection is not None:
-            # Get all stored embeddings from MongoDB
-            logger.debug("Querying MongoDB for stored embeddings")
-            stored_docs = list(embeddings_collection.find({}, {"embedding": 1, "code": 1, "function_name": 1}))
-            logger.debug(f"Retrieved {len(stored_docs)} embeddings from MongoDB for similarity comparison")
+#     try:
+#         if mongo_client is not None and embeddings_collection is not None:
+#             # Get all stored embeddings from MongoDB
+#             logger.debug("Querying MongoDB for stored embeddings")
+#             stored_docs = list(embeddings_collection.find({}, {"embedding": 1, "code": 1, "function_name": 1}))
+#             logger.debug(f"Retrieved {len(stored_docs)} embeddings from MongoDB for similarity comparison")
             
-            if not stored_docs:
-                logger.warning("No stored embeddings found in MongoDB")
-                return []
+#             if not stored_docs:
+#                 logger.warning("No stored embeddings found in MongoDB")
+#                 return []
             
-            # Calculate similarities
-            logger.debug("Calculating similarity scores between query and stored embeddings")
-            calc_start = datetime.now()
-            similarities = []
-            for doc in stored_docs:
-                similarity = compute_similarity(query_embedding, np.array(doc["embedding"]))
-                doc["similarity"] = similarity
-                similarities.append((similarity, doc))
+#             # Calculate similarities
+#             logger.debug("Calculating similarity scores between query and stored embeddings")
+#             calc_start = datetime.now()
+#             similarities = []
+#             for doc in stored_docs:
+#                 similarity = compute_similarity(query_embedding, np.array(doc["embedding"]))
+#                 doc["similarity"] = similarity
+#                 similarities.append((similarity, doc))
             
-            calc_time = (datetime.now() - calc_start).total_seconds()
-            logger.debug(f"Calculated {len(similarities)} similarity scores in {calc_time:.2f}s")
+#             calc_time = (datetime.now() - calc_start).total_seconds()
+#             logger.debug(f"Calculated {len(similarities)} similarity scores in {calc_time:.2f}s")
                 
-            # Sort by similarity and get top k
-            similarities.sort(reverse=True, key=lambda x: x[0])
-            result = [doc for _, doc in similarities[:top_k]]
+#             # Sort by similarity and get top k
+#             similarities.sort(reverse=True, key=lambda x: x[0])
+#             result = [doc for _, doc in similarities[:top_k]]
             
-            # Log result summary
-            if result:
-                top_results = [(doc.get("function_name", "Unknown"), doc.get("similarity", 0)) for doc in result]
-                result_summary = ", ".join([f"{name}({sim:.4f})" for name, sim in top_results])
-                logger.info(f"Top {len(result)} similar contexts: {result_summary}")
-            else:
-                logger.warning("No similar contexts found after filtering")
+#             # Log result summary
+#             if result:
+#                 top_results = [(doc.get("function_name", "Unknown"), doc.get("similarity", 0)) for doc in result]
+#                 result_summary = ", ".join([f"{name}({sim:.4f})" for name, sim in top_results])
+#                 logger.info(f"Top {len(result)} similar contexts: {result_summary}")
+#             else:
+#                 logger.warning("No similar contexts found after filtering")
             
-            elapsed = (datetime.now() - start_time).total_seconds()
-            logger.debug(f"Similar contexts retrieval completed in {elapsed:.2f}s")
-            return result
-        else:
-            logger.warning("MongoDB connection unavailable - cannot retrieve embeddings")
-            return []
-    except Exception as e:
-        elapsed = (datetime.now() - start_time).total_seconds()
-        logger.error(f"Error retrieving similar contexts after {elapsed:.2f}s: {e}")
-        logger.error(traceback.format_exc())
-        return []
+#             elapsed = (datetime.now() - start_time).total_seconds()
+#             logger.debug(f"Similar contexts retrieval completed in {elapsed:.2f}s")
+#             return result
+#         else:
+#             logger.warning("MongoDB connection unavailable - cannot retrieve embeddings")
+#             return []
+#     except Exception as e:
+#         elapsed = (datetime.now() - start_time).total_seconds()
+#         logger.error(f"Error retrieving similar contexts after {elapsed:.2f}s: {e}")
+#         logger.error(traceback.format_exc())
+#         return []
 
-def generate_feedback(student_code: str, ideal_code: str, similarity: float, 
-                    structure_analysis: Dict[str, Any], similar_contexts: List[Dict[str, Any]]) -> str:
-    """Generate feedback using Ollama."""
-    logger.info(f"Generating feedback with Ollama (similarity score: {similarity:.4f})")
-    start_time = datetime.now()
+# def generate_feedback(student_code: str, ideal_code: str, similarity: float, 
+#                     structure_analysis: Dict[str, Any], similar_contexts: List[Dict[str, Any]]) -> str:
+#     """Generate feedback using Ollama."""
+#     logger.info(f"Generating feedback with Ollama (similarity score: {similarity:.4f})")
+#     start_time = datetime.now()
     
-    try:
-        # Format the prompt
-        logger.debug("Formatting prompt for feedback generation")
-        prompt = f"""
-        You are an AI code reviewer evaluating a student's function implementation against an ideal solution. 
-        Your task is to provide constructive feedback on the correctness and quality of the student's code.
+#     try:
+#         # Format the prompt
+#         logger.debug("Formatting prompt for feedback generation")
+#         prompt = f"""
+#         You are an AI code reviewer evaluating a student's function implementation against an ideal solution. 
+#         Your task is to provide constructive feedback on the correctness and quality of the student's code.
 
-        ### Ideal Code:
-        ```python
-        {ideal_code}
-        ```
+#         ### Ideal Code:
+#         ```python
+#         {ideal_code}
+#         ```
 
-        ### Student Code:
-        ```python
-        {student_code}
-        ```
+#         ### Student Code:
+#         ```python
+#         {student_code}
+#         ```
 
-        ### Similarity Score: {similarity:.2f}
+#         ### Similarity Score: {similarity:.2f}
 
-        ### Structure Analysis:
-        {json.dumps(structure_analysis, indent=2)}
+#         ### Structure Analysis:
+#         {json.dumps(structure_analysis, indent=2)}
 
-        Based on this information, provide 2-3 paragraphs of constructive feedback for the student, 
-        highlighting strengths and areas for improvement.
-        """
+#         Based on this information, provide 2-3 paragraphs of constructive feedback for the student, 
+#         highlighting strengths and areas for improvement.
+#         """
         
-        # Call Ollama API
-        logger.debug(f"Calling Ollama API at {OLLAMA_BASE_URL}")
-        api_start = datetime.now()
-        response = requests.post(
-            f"{OLLAMA_BASE_URL}/api/chat",
-            json={
-                "model": "llama3.2:3b",
-                "messages": [
-                    {"role": "system", "content": "You are a helpful programming assistant."},
-                    {"role": "user", "content": prompt}
-                ],
-                "temperature": 0.7,
-                "stream": False
-            },
-            timeout=30
-        )
-        api_time = (datetime.now() - api_start).total_seconds()
-        logger.debug(f"Ollama API response received in {api_time:.2f}s (status: {response.status_code})")
+#         # Call Ollama API
+#         logger.debug(f"Calling Ollama API at {OLLAMA_BASE_URL}")
+#         api_start = datetime.now()
+#         response = requests.post(
+#             f"{OLLAMA_BASE_URL}/api/chat",
+#             json={
+#                 "model": "llama3.2:3b",
+#                 "messages": [
+#                     {"role": "system", "content": "You are a helpful programming assistant."},
+#                     {"role": "user", "content": prompt}
+#                 ],
+#                 "temperature": 0.7,
+#                 "stream": False
+#             },
+#             timeout=30
+#         )
+#         api_time = (datetime.now() - api_start).total_seconds()
+#         logger.debug(f"Ollama API response received in {api_time:.2f}s (status: {response.status_code})")
         
-        if response.status_code == 200:
-            result = response.json()
+#         if response.status_code == 200:
+#             result = response.json()
             
-            if "message" in result and "content" in result["message"]:
-                feedback = result["message"]["content"]
-                feedback_length = len(feedback)
-                logger.info(f"Generated feedback successfully ({feedback_length} chars)")
-                elapsed = (datetime.now() - start_time).total_seconds()
-                logger.debug(f"Feedback generation completed in {elapsed:.2f}s")
-                return feedback
-            elif "response" in result:
-                feedback = result["response"]
-                feedback_length = len(feedback)
-                logger.info(f"Generated feedback successfully ({feedback_length} chars)")
-                elapsed = (datetime.now() - start_time).total_seconds()
-                logger.debug(f"Feedback generation completed in {elapsed:.2f}s")
-                return feedback
-            else:
-                logger.error(f"Unexpected response format from Ollama: {result}")
-                return "Error generating feedback."
-        else:
-            error_msg = f"Error generating feedback: {response.status_code}"
-            logger.error(f"Ollama API error: {response.status_code}, {response.text}")
-            return error_msg
+#             if "message" in result and "content" in result["message"]:
+#                 feedback = result["message"]["content"]
+#                 feedback_length = len(feedback)
+#                 logger.info(f"Generated feedback successfully ({feedback_length} chars)")
+#                 elapsed = (datetime.now() - start_time).total_seconds()
+#                 logger.debug(f"Feedback generation completed in {elapsed:.2f}s")
+#                 return feedback
+#             elif "response" in result:
+#                 feedback = result["response"]
+#                 feedback_length = len(feedback)
+#                 logger.info(f"Generated feedback successfully ({feedback_length} chars)")
+#                 elapsed = (datetime.now() - start_time).total_seconds()
+#                 logger.debug(f"Feedback generation completed in {elapsed:.2f}s")
+#                 return feedback
+#             else:
+#                 logger.error(f"Unexpected response format from Ollama: {result}")
+#                 return "Error generating feedback."
+#         else:
+#             error_msg = f"Error generating feedback: {response.status_code}"
+#             logger.error(f"Ollama API error: {response.status_code}, {response.text}")
+#             return error_msg
             
-    except Exception as e:
-        elapsed = (datetime.now() - start_time).total_seconds()
-        logger.error(f"Error generating feedback after {elapsed:.2f}s: {e}")
-        logger.error(traceback.format_exc())
-        return f"Error generating feedback: {str(e)}"
+#     except Exception as e:
+#         elapsed = (datetime.now() - start_time).total_seconds()
+#         logger.error(f"Error generating feedback after {elapsed:.2f}s: {e}")
+#         logger.error(traceback.format_exc())
+#         return f"Error generating feedback: {str(e)}"
 
 @app.post("/evaluate")
 async def evaluate(
@@ -1240,7 +1246,7 @@ async def main_page(request: Request):
     return templates.TemplateResponse("index.html", {"request": request})
 
 @app.post("/api/process-qa")
-async def process_qa():
+async def process_qa(request: Request):
     """Process Q&A evaluation request."""
     try:
         # Parse the JSON data from the request
